`tile-layer` directory consolidation
